+++ metadata.rb
@@ -4,9 +4,10 @@ maintainer_email 'radek@opsline.com'
 license          'All rights reserved'
 description      'Installs/Configures opsline-rails-app'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.5.0'
+version          '0.5.2'
 
 depends 'artifact'
 depends 'ruby_build'
 depends 'rbenv', '~> 0.7.2'
 depends 'rvm', '~> 0.9.2'
+depends 'logrotate'