+++ recipes/deploy.rb
@@ -5,7 +5,11 @@ end
 # install all configured rails apps
 node['opsline-rails-app']['apps'].each do |app_id|
 
-  app_data = data_bag_item(node['opsline-rails-app']['databag'], app_id)
+  if node['opsline-rails-app']['encrypted_databag']
+    app_data = Chef::EncryptedDataBagItem.load(node['opsline-rails-app']['databag'], app_id).to_hash
+  else
+    app_data = data_bag_item(node['opsline-rails-app']['databag'], app_id)
+  end
 
   # get app name
   if app_data.has_key?('name')
@@ -21,24 +25,28 @@ node['opsline-rails-app']['apps'].each do |app_id|
   unless app_data.has_key?('deploy_to')
     app_data['deploy_to'] = "#{node['opsline-rails-app']['apps_root']}/#{app_name}"
   end
+  unless app_data.has_key?('package_type')
+    app_data['package_type'] = 'tar.gz'
+  end
   unless app_data.has_key?('artifact_name')
     app_data['artifact_name'] = app_name
   end
   unless app_data.has_key?('artifact_location')
     app_data['artifact_location'] = "s3://s3.amazonaws.com/#{node['opsline-rails-app']['s3_bucket']}/#{app_data['artifact_name']}"
   end
-  unless app_data.has_key?('package_type')
-    app_data['package_type'] = 'tar.gz'
+  app_data['version'] = get_env_value(app_data['version'])
+  app_data['artifact'] = "#{app_data['artifact_name']}-#{app_data['version']}.#{app_data['package_type']}"
+  if app_data.has_key?('jenkins_job_name')
+    app_data['artifact'] = "jobs/#{app_data['jenkins_job_name']}/#{app_data['version']}/#{app_data['artifact']}"
   end
   unless app_data.has_key?('container')
     app_data['container'] = 'unicorn'
   end
   if app_data.has_key?('container_parameters')
-    container_parameters = get_env_value(app_data['container_parameters'])
+    app_data['container_parameters'] = get_env_value(app_data['container_parameters'])
   else
-    container_parameters = {}
+    app_data['container_parameters'] = {}
   end
-  artifact_version = get_env_value(app_data['version'])
 
   # create app directory
   directory app_data['deploy_to'] do
@@ -49,20 +57,20 @@ node['opsline-rails-app']['apps'].each do |app_id|
   end
 
   # environment variables hash
-  env_dict = {}
-  env_dict['RACK_ENV'] = node.chef_environment
-  env_dict['RAILS_ENV'] = node.chef_environment
-  env_dict['HOME'] = "/home/#{node['opsline-rails-app']['owner']}"
-  env_dict['PATH'] = "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
+  app_data['app_env'] = {}
+  app_data['app_env']['RACK_ENV'] = node.chef_environment
+  app_data['app_env']['RAILS_ENV'] = node.chef_environment
+  app_data['app_env']['HOME'] = "/home/#{node['opsline-rails-app']['owner']}"
+  app_data['app_env']['PATH'] = "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin"
   if node['opsline-rails-app']['ruby']['provider'] == 'rvm'
-    env_dict['PATH'] = "/home/#{node['opsline-rails-app']['owner']}/.rvm/bin:#{env_dict['PATH']}"
+    app_data['app_env']['PATH'] = "/home/#{node['opsline-rails-app']['owner']}/.rvm/bin:#{app_data['app_env']['PATH']}"
   elsif node['opsline-rails-app']['ruby']['provider'] == 'rbenv'
-    env_dict['PATH'] = "/home/#{node['opsline-rails-app']['owner']}/.rbenv/shims:/home/#{node['opsline-rails-app']['owner']}/.rbenv/bin:#{env_dict['PATH']}"
+    app_data['app_env']['PATH'] = "/home/#{node['opsline-rails-app']['owner']}/.rbenv/shims:/home/#{node['opsline-rails-app']['owner']}/.rbenv/bin:#{app_data['app_env']['PATH']}"
   end
 
   # merge environment from data bag (if there)
   if app_data.has_key?('environment')
-    env_dict.merge!(get_env_value(app_data['environment']))
+    app_data['app_env'].merge!(get_env_value(app_data['environment']))
   end
 
   # install pre-requisite packages
@@ -74,30 +82,58 @@ node['opsline-rails-app']['apps'].each do |app_id|
     end
   end
 
+  # configure logrotate
+  unless app_data.has_key?('logrotate')
+    logrotate = true
+  else
+    logrotate = to_bool(app_data['logrotate'])
+  end
+  if logrotate
+    unless app_data.has_key?('logrotate_days')
+      days = 7
+    else
+      days = app_data['logrotate_days'].to_i
+    end
+    logrotate_app app_name do
+      cookbook 'logrotate'
+      path "#{app_data['deploy_to']}/shared/log/*.log"
+      options ['copytruncate', 'missingok', 'compress', 'notifempty', 'delaycompress']
+      frequency 'daily'
+      rotate days
+    end
+  end
+
+  app_data['shared_directories'] = %w{ bundle config log pids sockets system }
+  app_data['symlinks'] = {'log' => 'log'}
+  # additional shared_subdirectories
+  if app_data.has_key?('shared_subdirectories')
+    app_data['shared_subdirectories'].each do |dir|
+      app_data['shared_directories'] << dir
+      app_data['symlinks'][dir] = dir
+    end
+  end
+
   services_to_restart = []
   pids_to_signal = []
-
-
+  restart_after_deploy = false
 
   # deploy artifact
   artifact_deploy app_name do
-    version artifact_version
-    artifact_location "#{app_data['artifact_location']}/#{app_data['artifact_name']}-#{artifact_version}.#{app_data['package_type']}"
+    version app_data['version']
+    artifact_location "#{app_data['artifact_location']}/#{app_data['artifact']}"
     deploy_to app_data['deploy_to']
     owner node['opsline-rails-app']['owner']
     group node['opsline-rails-app']['owner']
-    environment env_dict
-    shared_directories %w{ bundle config log pids sockets system }
-    symlinks({
-      'log' => 'log'
-    })
+    environment app_data['app_env']
+    shared_directories app_data['shared_directories']
+    symlinks(app_data['symlinks'])
     action :deploy
     keep 3
     force false
 
     # remove log directory before linking
     before_symlink Proc.new {
-      directory "#{app_data['deploy_to']}/releases/#{artifact_version}/log" do
+      directory "#{app_data['deploy_to']}/releases/#{app_data['version']}/log" do
         action :delete
         recursive true
       end
@@ -108,23 +144,23 @@ node['opsline-rails-app']['apps'].each do |app_id|
 
       # UNICORN
       if app_data['container'] == 'unicorn'
-        unless container_parameters.has_key?('timeout')
-          container_parameters['timeout'] = node['opsline-rails-app']['unicorn']['timeout']
+        unless app_data['container_parameters'].has_key?('timeout')
+          app_data['container_parameters']['timeout'] = node['opsline-rails-app']['unicorn']['timeout']
         end
-        unless container_parameters.has_key?('worker_processes')
-          container_parameters['worker_processes'] = node['opsline-rails-app']['unicorn']['worker_processes']
+        unless app_data['container_parameters'].has_key?('worker_processes')
+          app_data['container_parameters']['worker_processes'] = node['opsline-rails-app']['unicorn']['worker_processes']
         end
-        unless container_parameters.has_key?('frontend')
-          container_parameters['frontend'] = 'nginx'
+        unless app_data['container_parameters'].has_key?('frontend')
+          app_data['container_parameters']['frontend'] = 'nginx'
         end
-        unless container_parameters.has_key?('frontend_port')
-          container_parameters['frontend_port'] = '8080'
+        unless app_data['container_parameters'].has_key?('frontend_port')
+          app_data['container_parameters']['frontend_port'] = '8080'
         end
         unicorn_config = "#{app_data['deploy_to']}/shared/config/unicorn.rb"
 
         env_dir app_data do
           deploy_to app_data['deploy_to']
-          variables env_dict
+          variables app_data['app_env']
           owner node['opsline-rails-app']['owner']
           group node['opsline-rails-app']['owner']
           notifies [:restart, "service[#{app_name}]"]
@@ -142,8 +178,8 @@ node['opsline-rails-app']['apps'].each do |app_id|
             :app_name => app_name,
             :user => node['opsline-rails-app']['owner'],
             :deploy_to => app_data['deploy_to'],
-            :environment => env_dict,
-            :parameters => container_parameters
+            :environment => app_data['app_env'],
+            :parameters => app_data['container_parameters']
           })
         end
 
@@ -174,71 +210,34 @@ node['opsline-rails-app']['apps'].each do |app_id|
           services_to_restart << [app_name, Chef::Provider::Service::Upstart]
         end
 
-        if container_parameters['frontend'] == 'nginx'
-          service 'nginx' do
-            action :nothing
-          end
+        if app_data['container_parameters']['frontend'] == 'nginx'
           services_to_restart << ['nginx', Chef::Provider::Service::Init]
-          directory '/etc/nginx'
-          directory '/etc/nginx/sites-available'
-          directory '/etc/nginx/sites-enabled'
-          template "/etc/nginx/sites-available/#{app_name}" do
-            source 'nginx.unicorn.conf.erb'
-            cookbook 'opsline-rails-app'
-            owner 'root'
-            group 'root'
-            mode 0644
-            action :create
-            notifies :restart, resources(:service => 'nginx'), :delayed
-            variables({
-              :app_name => app_name,
-              :deploy_to => app_data['deploy_to'],
-              :port => container_parameters['frontend_port'],
-              :server_name => "#{app_name}.#{node.domain}"
-            })
-          end
-          link "#{app_name} nginx site" do
-            target_file "/etc/nginx/sites-enabled/#{app_name}"
-            to "/etc/nginx/sites-available/#{app_name}"
+          app_data['container_parameters']['upstream_sockets'] = [
+            "#{app_data['deploy_to']}/shared/sockets/#{app_name}.socket"
+          ]
+
+          nginx_app_config "nginx config for #{app_name}" do
+            app_name app_name
+            app_data app_data
           end
         end
 
       # PASSENGER
       elsif app_data['container'] == 'passenger'
-        unless container_parameters.has_key?('frontend')
-          container_parameters['frontend'] = 'nginx'
+        unless app_data['container_parameters'].has_key?('frontend')
+          app_data['container_parameters']['frontend'] = 'nginx'
         end
-        unless container_parameters.has_key?('frontend_port')
-          container_parameters['frontend_port'] = '8080'
+        unless app_data['container_parameters'].has_key?('frontend_port')
+          app_data['container_parameters']['frontend_port'] = '8080'
         end
 
-        if container_parameters['frontend'] == 'nginx'
-          service 'nginx' do
-            action :nothing
-          end
+        if app_data['container_parameters']['frontend'] == 'nginx'
           services_to_restart << ['nginx', Chef::Provider::Service::Init]
-          directory '/etc/nginx'
-          directory '/etc/nginx/sites-available'
-          directory '/etc/nginx/sites-enabled'
-          template "/etc/nginx/sites-available/#{app_name}" do
-            source 'nginx.passenger.conf.erb'
-            cookbook 'opsline-rails-app'
-            owner 'root'
-            group 'root'
-            mode 0644
-            action :create
-            notifies :restart, resources(:service => 'nginx'), :delayed
-            variables({
-              :app_name => app_name,
-              :deploy_to => app_data['deploy_to'],
-              :port => container_parameters['frontend_port'],
-              :server_name => "#{app_name}.#{node.domain}",
-              :env => env_dict
-            })
-          end
-          link "#{app_name} nginx site" do
-            target_file "/etc/nginx/sites-enabled/#{app_name}"
-            to "/etc/nginx/sites-available/#{app_name}"
+
+          nginx_app_config "nginx config for #{app_name}" do
+            app_name app_name
+            app_data app_data
+            template 'nginx.passenger.conf.erb'
           end
         end
 
@@ -246,7 +245,7 @@ node['opsline-rails-app']['apps'].each do |app_id|
       elsif app_data['container'] == 'rack'
         env_dir app_data do
           deploy_to app_data['deploy_to']
-          variables env_dict
+          variables app_data['app_env']
           owner node['opsline-rails-app']['owner']
           group node['opsline-rails-app']['owner']
           notifies [:restart, "service[#{app_name}]"]
@@ -276,16 +275,16 @@ node['opsline-rails-app']['apps'].each do |app_id|
 
       # WORKER
       elsif app_data['container'] == 'worker'
-        unless container_parameters.has_key?('number_of_workers')
-          container_parameters['number_of_workers'] = 4
+        unless app_data['container_parameters'].has_key?('number_of_workers')
+          app_data['container_parameters']['number_of_workers'] = 4
         end
-        unless container_parameters.has_key?('bundle_exec_command')
-          container_parameters['bundle_exec_command'] = ''
+        unless app_data['container_parameters'].has_key?('bundle_exec_command')
+          app_data['container_parameters']['bundle_exec_command'] = ''
         end
 
         env_dir app_data do
           deploy_to app_data['deploy_to']
-          variables env_dict
+          variables app_data['app_env']
           owner node['opsline-rails-app']['owner']
           group node['opsline-rails-app']['owner']
           notifies [:restart, "service[#{app_name}]"]
@@ -304,7 +303,7 @@ node['opsline-rails-app']['apps'].each do |app_id|
             :user => node['opsline-rails-app']['owner'],
             :group => node['opsline-rails-app']['owner'],
             :deploy_to => app_data['deploy_to'],
-            :exec_command => container_parameters['bundle_exec_command']
+            :exec_command => app_data['container_parameters']['bundle_exec_command']
           })
         end
         template "/etc/init/#{app_name}.conf" do
@@ -320,7 +319,7 @@ node['opsline-rails-app']['apps'].each do |app_id|
             :user => node['opsline-rails-app']['owner'],
             :group => node['opsline-rails-app']['owner'],
             :deploy_to => app_data['deploy_to'],
-            :number_of_workers => container_parameters['number_of_workers']
+            :number_of_workers => app_data['container_parameters']['number_of_workers']
           })
         end
         service app_name do
@@ -334,19 +333,15 @@ node['opsline-rails-app']['apps'].each do |app_id|
 
     # restart proc
     restart Proc.new {
-      services_to_restart.each do |service_to_restart|
-        service service_to_restart[0] do
-          provider service_to_restart[1]
-          action :restart
-        end
-      end
-      pids_to_signal.each do |pid_to_signal|
-        pid = pid_to_signal[0]
-        signal = pid_to_signal[1]
-        execute "signal #{signal} #{pid}" do
-          user 'root'
-          command "kill -#{signal} #{pid}"
-          action :run
+      restart_after_deploy = true
+    }
+
+    # after_deploy proc
+    after_deploy Proc.new {
+      if restart_after_deploy
+        restart_services "restart #{app_name} services" do
+          services_to_restart services_to_restart
+          pids_to_signal pids_to_signal
         end
       end
     }