+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # CHANGELOG for opsline-mysql
 
+## 0.5.2
+* nginx setup rewrite
+
+## 0.5.1
+* Adding logrotate
+
 ## 0.5.0
-# Adding worker container
+* Adding worker container
 * Ability to install multiple versions of ruby via rbenv
 * Support for rvm
+* Using encrypted data bags
 
 ## 0.1.0:
 * Initial release